backend/requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.32
pydantic>=2.10
pydantic-settings>=2.6
sqlalchemy[asyncio]>=2.0
asyncpg>=0.30
python-jose[cryptography]>=3.3
pytest>=8.3
httpx>=0.28
